Removida requisição de troca de senha da tela de troca de senha
