+++ src/pages/ChangePassword/index.jsx
@@ -1,20 +1,20 @@
 import React, { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faArrowLeft, faArrowLeftLong, faArrowRotateBack, faCheck } from '@fortawesome/free-solid-svg-icons'
-import { Button, Col, Form, Row} from 'react-bootstrap';
-import { AuthAPI } from "../../api/auth-api";
-import { AUTH_DEBUG, BASE_URL, HttpStatus } from "../../api/default";
-import { ToastContainer, toast } from 'react-toastify';
+import { faArrowLeftLong, faCheck } from '@fortawesome/free-solid-svg-icons'
+import { Form } from 'react-bootstrap';
+import { HttpStatus } from "../../api/default";
+import { toast } from 'react-toastify';
+import { PasswordAPI } from "../../api/password"
 import 'react-toastify/dist/ReactToastify.css';
 
 // Style
 import './style.css'
 import { useAuthContext } from "../../contexts/AuthContext";
-import { HttpResponse } from "../CreateCourse/api";
 
 
-  
+
+
 
 const ChangePasswordPage = () => {
     const navigate = useNavigate();
@@ -24,7 +24,7 @@ const ChangePasswordPage = () => {
     const [formValues, setFormValues] = useState(intialValues);
     const [formErrors, setFormErrors] = useState({});
     const [isSubmit, setIsSubmit] = useState(false)
-    const [show, setShow] = useState(false);
+
     const notifySuccess = (texto) => toast.success(texto, {
         position: "top-right",
         autoClose: 5000,
@@ -34,8 +34,8 @@ const ChangePasswordPage = () => {
         draggable: true,
         progress: undefined,
         theme: "light",
-      });
-      const notifyError = (texto) => toast.error(texto, {
+    });
+    const notifyError = (texto) => toast.error(texto, {
         position: "top-right",
         autoClose: 5000,
         hideProgressBar: false,
@@ -44,7 +44,7 @@ const ChangePasswordPage = () => {
         draggable: true,
         progress: undefined,
         theme: "light",
-      });
+    });
 
     var errorsC = 0;
 
@@ -64,44 +64,16 @@ const ChangePasswordPage = () => {
         setFormErrors(validate(formValues));
         setIsSubmit(true);
         if (errorsC == 0) {
-           // console.log("Requisição realizada.")
-            //console.log(formValues)
-            const response = await fetchChange()
-           // console.log(response)
+            const response = await PasswordAPI.fetchChange(formValues.oldPassword, formValues.newPassword, token);
             if (response.status == HttpStatus.OK) {
                 notifySuccess("Senha alterada com sucesso.");
-            }else{
-                notifyError("Falha em alterar senha.");
+            } else {
+                notifyError(response.data.error);
             }
 
         }
     }
 
-    const fetchChange = async () => {
-        const url = `${BASE_URL}/change-password`
-        try {
-            const options = {
-                method: 'PUT',
-                credentials: 'include',
-                headers: {
-                    'Content-Type': 'application/json',
-                    Accept: 'application/json', 
-                    'jwt': token
-                },
-                body: JSON.stringify({ old_password: formValues.oldPassword, new_password: formValues.newPassword })
-            }
-
-            const response = await fetch(url, options);
-            if (response.ok) {
-                const data = await response.json();
-                AUTH_DEBUG && console.log("AuthAPI::ChangePassword(): ", data.token);
-                return new HttpResponse(HttpStatus.OK, data);
-            } else throw new Error("Error on ChangePassword()");
-        } catch (error) {
-            console.warn(error)
-            return new HttpResponse(HttpStatus.ERROR, null);
-        }
-    }
 
     const validate = (values) => {
         const errors = {};
@@ -139,7 +111,7 @@ const ChangePasswordPage = () => {
                                 <img style={{ width: "11em" }} onClick={() => { navigate("/") }} src="https://i.ibb.co/r3QPmSt/logo.png" alt="logo" border="0" />
                             </div>
                             <div className="col d-flex justify-content-end">
-                                <p className="mt-3"><Link className="fw-bold link-limpo" style={{ color: "#1dbfb0" }} to="/perfil"><FontAwesomeIcon icon={faArrowLeftLong} className="me-1"/> Voltar!</Link></p>
+                                <p className="mt-3"><Link className="fw-bold link-limpo" style={{ color: "#1dbfb0" }} to="/perfil"><FontAwesomeIcon icon={faArrowLeftLong} className="me-1" /> Voltar!</Link></p>
                             </div>
                         </div>
                     </div>
